Updated game_loop for edge cases
Included various stopping points for if the player or dealer gets 21.

FINAL gameplay problem is the ace mechanic. Need to figure out where to incorporate an ace being an 11 or a 1 depending on hand_value. Will most likely need to make these corrections within the hand_value function and not the card_value function